Add option to create a PointerBarrier in front of the bar
This makes it impossible for the pointer to enter the bar area.
Effectively, the pointer can only move in the area defined by
m->w{x,y,w,h}.

It is important, though, to make that barrier transparent in one
direction. The pointer must be able to *leave* the bar area because it
might slip into it somehow (X bugs, Xinerama, ...).

Requires XFIXES 5 or above:

http://cgit.freedesktop.org/xorg/proto/fixesproto/commit/?id=9760b4bdd1f9fdd6a33b9f876c4a835ed969aa84

See also:

http://who-t.blogspot.de/2012/12/whats-new-in-xi-23-pointer-barrier.html