Use "nmaster" only if there's more than 3 clients
I find "nmaster = 2" very comfortable -- but only if there's enough
windows on screen. For two windows, it's awkward. For three windows,
it's unbalanced because the very first window sits alone in the slave
column and it is larger than the second window. For four or more
windows, "nmaster = 2" is actually useful.

This is an experimental change. I'll have to try this for a little
while. Eventually, I might turn it into a proper configuration option
(config.h).