Read virtual monitors from property on root window
You can now do this:

    $ xprop -root \
        -f DWM_VIRTUAL_MONITORS 8s \
        -set DWM_VIRTUAL_MONITORS '1024x768+0+0;800x600+1280+20;'

To return to "xinerama only", simply remove the property:

    $ xprop -root -remove DWM_VIRTUAL_MONITORS

The format of one geometry string is as follows:

    "<width>{xX}<height>{+}<xoffset>{+}<yoffset>"

In other words, you must provide width, height, xoffset and yoffset.
Also, xoffset and yoffset must not be negative.

Each geometry string must be terminated with a ";".

(It would have been easier for dwm to use an integer array property.
However, there is no easy way to set such a property -- xprop cannot do
this. I would have needed to write a new tool for this job.)